feat: Refactor and add optimisation logic
Refactor Option methods names to be more understandable.
Add new filters on the final step of handling option changes in OptionsAccessHandler.applySaveModifiedOptionsThenRunReloaders, taking into account the options that might have been modified with pending values equal to actual value. In that case, the appliance do not complete, and the reloaders and storages are not added to the reload/save filters. A comparison is done between the final lists and the filters, and all members that are not in the filters are removed.